Move get_pub_date() to avoid cyclic import.
